fix: インタラクションハンドラーも分離した
